backend/requirements.txt
Django==5.2
Pillow==11.2.1
djangorestframework==3.16.0
drf-yasg==1.21.10
django-cors-headers==4.7.0
psycopg2-binary==2.9.10